Remove some mappers, reuse existing objects, static methods instead of DI
