+++ graylog2-server/src/main/java/org/graylog/plugins/views/search/rest/scriptingapi/request/SearchRequestSpec.java
@@ -18,13 +18,16 @@ package org.graylog.plugins.views.search.rest.scriptingapi.request;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.google.common.base.Strings;
+import org.graylog.plugins.views.search.rest.scriptingapi.parsing.TimerangeParser;
 import org.graylog2.plugin.indexer.searches.timeranges.RelativeRange;
 import org.graylog2.plugin.indexer.searches.timeranges.TimeRange;
 
 import javax.validation.Valid;
 import javax.validation.constraints.NotEmpty;
+import javax.ws.rs.BadRequestException;
 import java.util.List;
 import java.util.Set;
+import java.util.stream.Collectors;
 
 public record SearchRequestSpec(@JsonProperty("query") String queryString,
                                 @JsonProperty("streams") Set<String> streams,
@@ -48,5 +51,25 @@ public record SearchRequestSpec(@JsonProperty("query") String queryString,
 
     public boolean hasCustomSort() {
         return metrics().stream().anyMatch(m -> m.sort() != null);
+    }
+
+    public static SearchRequestSpec fromSimpleParams(String query, Set<String> streams, String timerangeKeyword, List<String> groups, List<String> metrics) {
+        if (groups == null || groups.isEmpty()) {
+            throw new BadRequestException("At least one grouping has to be provided!");
+        }
+        if (metrics == null || metrics.isEmpty()) {
+            metrics = List.of("count:");
+        }
+        if (!metrics.stream().allMatch(m -> m.contains(":"))) {
+            throw new BadRequestException("All metrics need to be defined as \"function\":\"field_name\"");
+        }
+
+        return new SearchRequestSpec(
+                query,
+                streams,
+                TimerangeParser.parseTimeRange(timerangeKeyword),
+                groups.stream().map(Grouping::new).collect(Collectors.toList()),
+                metrics.stream().map(Metric::parseMetric).collect(Collectors.toList())
+        );
     }
 }
